explicitly track available resources to avoid racing the idle closer
Instead of measuring available resources directly from the channel, use
an explicit counter in order to avoid short-term blips when the idle
closer grabs a resource from the pool.

Note that this only really affects tests since in the real world the race
is likely to be infinitesimal, but rather than mark the tests as flaky this
seemed like the better solution.