pools: option to prefill
Background: this change allows one to create prefilled resource
pools. This is useful when traffics suddenly shift while the
pool is still empty. This causes a thundering herd of Open requests
that can cause outage.

There is a proposal to rewrite this pool to natively accommodate
this feature. This change is a stand-in until that effort is
completed.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>